Arreglos y cambios
GameObject tiene SceneNode y una referencia al transform
Mejorado código de varios componentes

+++ MotorGaia/Src/Gaia/src/RenderSystem.cpp
@@ -1,17 +1,30 @@
 #include "RenderSystem.h"
 
-RenderSystem::RenderSystem() {}
-
-RenderSystem::~RenderSystem() {}
-
-void RenderSystem::setup(Ogre::Root* root)
+RenderSystem::RenderSystem()
 {
-	sceneManager = root->createSceneManager();
+
 }
 
-void RenderSystem::shutDown(Ogre::Root* root)
+RenderSystem::~RenderSystem()
 {
-	root->destroySceneManager(sceneManager);
+
+}
+
+void RenderSystem::setup(Ogre::Root* _root)
+{
+	root = _root;
+	sceneManager = _root->createSceneManager();
+}
+
+void RenderSystem::render()
+{
+	root->renderOneFrame();
+}
+
+void RenderSystem::shutDown()
+{
+	if (root != nullptr)
+		root->destroySceneManager(sceneManager);
 }
 
 Ogre::SceneManager* RenderSystem::getSceneManager()
